Reorganize and change subcommand name.
